+++ webapp/lib/queries.py
@@ -113,9 +113,10 @@ class VFQueries():
                 return df.unstack(level=0, fill_value=0)
             else:
                 df.columns = [col for col in df["count"].columns.values]
-                for taxid in ids:
-                    if taxid not in df.columns:
-                        df[taxid] = 0
+                if search_on == "taxid":
+                    for taxid in ids:
+                        if taxid not in df.columns:
+                            df[taxid] = 0
 
         elif indexing == "seqid":
             if plasmids is not None: