+++ webapp/views/fam.py
@@ -47,9 +47,7 @@ def tab_gen_profile_tree(db, main_series, header, intersect):
     e_tree = EteTree(tree)
     e_tree.rename_leaves(ref_names)
 
-    face_params = {"color": EteTree.RED}
-    e_tree.add_column(SimpleColorColumn.fromSeries(main_series,
-                                                   header=header, face_params=face_params))
+    e_tree.add_column(SimpleColorColumn.fromSeries(main_series, header=header))
 
     for og in df_og_count:
         og_serie = df_og_count[og]
@@ -125,7 +123,7 @@ class FamBaseView(View):
             self.db, getattr(hit_counts, self.object_column),
             self.format_entry(entry_id), orthogroups)
         asset_path = f"/temp/fam_tree_{entry_id}.svg"
-        path = settings.BASE_DIR + "/assets/" + asset_path
+        path = settings.BASE_DIR + "/assets" + asset_path
         e_tree.render(path, dpi=500)
 
         info = {self.colname_to_header(key): infos[key]